Moved utility classes into utils package
