+++ integration/spark/src/main/java/io/openlineage/spark/agent/OpenLineageSparkListener.java
@@ -7,8 +7,8 @@ import io.openlineage.spark.agent.client.OpenLineageClient;
 import io.openlineage.spark.agent.lifecycle.ContextFactory;
 import io.openlineage.spark.agent.lifecycle.ExecutionContext;
 import io.openlineage.spark.agent.lifecycle.SparkSQLExecutionContext;
-import io.openlineage.spark.agent.lifecycle.plan.ScalaConversionUtils;
 import io.openlineage.spark.agent.transformers.PairRDDFunctionsTransformer;
+import io.openlineage.spark.agent.util.ScalaConversionUtils;
 import java.lang.reflect.Field;
 import java.net.URISyntaxException;
 import java.time.ZonedDateTime;
