+++ integration/spark/src/main/java/io/openlineage/spark/agent/lifecycle/ContextFactory.java
@@ -7,7 +7,6 @@ import io.openlineage.spark.agent.lifecycle.plan.CommandPlanVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.DatasetSourceVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.LogicalRDDVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.LogicalRelationVisitor;
-
 import java.util.ArrayList;
 import java.util.List;
 import lombok.AllArgsConstructor;
