+++ integration/spark/src/main/java/io/openlineage/spark/agent/lifecycle/RddExecutionContext.java
@@ -9,8 +9,8 @@ import io.openlineage.spark.agent.client.DatasetParser;
 import io.openlineage.spark.agent.client.DatasetParser.DatasetParseResult;
 import io.openlineage.spark.agent.client.OpenLineageClient;
 import io.openlineage.spark.agent.facets.ErrorFacet;
-import io.openlineage.spark.agent.lifecycle.plan.PlanUtils;
-import io.openlineage.spark.agent.lifecycle.plan.ScalaConversionUtils;
+import io.openlineage.spark.agent.util.PlanUtils;
+import io.openlineage.spark.agent.util.ScalaConversionUtils;
 import java.io.IOException;
 import java.lang.reflect.Field;
 import java.net.URI;
