+++ integration/spark/src/main/java/io/openlineage/spark/agent/lifecycle/plan/AppendDataVisitor.java
@@ -2,6 +2,8 @@ package io.openlineage.spark.agent.lifecycle.plan;
 
 import io.openlineage.client.OpenLineage;
 import java.util.List;
+
+import io.openlineage.spark.agent.util.PlanUtils;
 import org.apache.spark.sql.catalyst.plans.logical.AppendData;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
 import scala.PartialFunction;
