+++ integration/spark/src/main/java/io/openlineage/spark/agent/lifecycle/plan/BigQueryNodeVisitor.java
@@ -8,6 +8,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
+
+import io.openlineage.spark.agent.util.PlanUtils;
 import org.apache.spark.sql.SQLContext;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
 import org.apache.spark.sql.execution.datasources.LogicalRelation;
